requirements.txt
numpy==1.19.1
Pillow==9.0.0
